chore(gitignore): ignore haskell template pattern
The main purpose of this fix is to ignore the `dist-newstyle` directory that cabal routinely creates from git.
For this reason, I have added patterns conventionally used in Haskell projects.
For compatibility reasons,
I have not deleted existing patterns,
but to make it clearer when there are duplicates,
I have applied dictionary order sorting.